CS6015: Malloc requested fixes

+++ Spring2026/CS6013/Week9/MallocReplacement/src/MyMalloc.h
@@ -10,7 +10,7 @@
 
 class MyMalloc {
 public:
-    void* allocate(int bytesToAllocate);
+    void* allocate(size_t bytesToAllocate);
     void deallocate(void* ptr);
 
     MyMalloc();
